סיום עבודה על אלגוריתמיקת שמירת התוצאות - מחלקה לשמירת המיגע - מערך שמאחסן בתוכו ארועים מהמחלקה
