+++ Press_Competition/T10-BTN_Config.ino
@@ -1,3 +1,4 @@
+#include "T26-History_Results.h"
 #define BTN D4
 
 ResultData* resultsHistory[10];
